tools/player: switch text_last_{frame_index,duration} type from int to int64_t
duration is also adjusted to int64_t because it is derived from
p->duration (which is int64_t) and then assigned to text_last_duration.

p->frame_index is already int64_t so it makes sense to align
text_last_frame_index with it.

+++ ngl-tools/player.h
@@ -64,8 +64,8 @@ struct player {
     int64_t lasthover;
     int mouse_down;
     int fullscreen;
-    int text_last_frame_index;
-    int text_last_duration;
+    int64_t text_last_frame_index;
+    int64_t text_last_duration;
     struct ngl_node *pgbar_opacity_node;
     struct ngl_node *pgbar_text_node;
     struct ngl_node *pgbar_duration_node;
